Color support and new data structure
Changed a few small things, and one big thing. The way the LUA code processes the data sent to it is now done with less code, and the code is made more readable. The ability to change the color of what you are drawing on the screen is now possible.
Also, drawType has been renamed to mode, since it does much more than drawing now. The reason for changing the way that data is processed can be seen in the way data is sent. Data can be sent with strings as indexes, and the frontend in Stormworks will turn the data into an array of all the commands sent from the file. This data will be rendered as one image.